+++ autoprotocol/util.py
@@ -102,6 +102,39 @@ def quad_num_to_ind(q, human=False):
         raise ValueError("Invalid quadrant number.")
 
 
+def check_valid_origin(origin, plate_type, xfer_axis):
+    # Checks if selected well is a valid origin destination for the given plate
+    # Assumption: SBS formatted plates and 96-tip layout
+    robotized_origin = plate_type.robotize(origin)
+    well_count = plate_type.well_count
+    col_count = plate_type.col_count
+    row_count = plate_type.well_count // col_count
+
+    if well_count == 96:
+        if xfer_axis == "row":
+            if (robotized_origin % col_count) != 0:
+                raise ValueError("For row transfers, origin"
+                                 "has to be specified within the left "
+                                 "column.")
+        else:
+            if robotized_origin > col_count or robotized_origin < 0:
+                raise ValueError("For column transfers, origin "
+                                 "has to be specified within the top "
+                                 "column.")
+    elif well_count == 384:
+        if xfer_axis == "row":
+            if (robotized_origin % col_count) not in [0, 1]:
+                raise ValueError("For row transfers, origin"
+                                 "has to be specified within the left "
+                                 "column.")
+        else:
+            if robotized_origin > col_count*2 or robotized_origin < 0:
+                raise ValueError("For column transfers, origin "
+                                 "has to be specified within the top "
+                                 "column.")
+    else:
+        raise RuntimeError("Unsupported plate type for checking origin.")
+
 class make_dottable_dict(dict):
     '''Enable dictionaries to be accessed using dot notation instead of bracket
     notation.